Deprecate torch.triangular_solve
There is a use of `at::triangular_solve_out` in the file
`torch/csrc/jit/tensorexpr/external_functions.cpp` that I have not dared
to move to `at::linalg_triangular_solve_out`.

**BC-breaking note:**

This PR deprecates the `torch.triangular_solve` function in favor of
`torch.linalg.triangular_solve`. An upgarde guide is added to the
documentation for `torch.triangular_solve`.

Note that it DOES NOT remove `torch.triangular_solve`, but
`torch.triangular_solve` will be removed in a future PyTorch release.

ghstack-source-id: 8b46762f83d16b24e55596d820e0deb69cace8ed
Pull Request resolved: https://github.com/pytorch/pytorch/pull/63570